requirements.txt
torch==2.5.0.dev20240702
torchaudio==2.4.0.dev20240702
torchvision==0.20.0.dev20240702
transformers==4.42.3
huggingface-hub==0.23.4
av==12.2.0
accelerate==0.31.0
pillow==10.4.0
sentencepiece==0.2.0
protobuf==5.27.2
numpy==1.26.4